+++ src/Themes/constants.js
@@ -1,0 +1,12 @@
+import basicSales from './basicSales';
+
+export const THEMES_TYPE = {
+  BASIC_SALES: 0,
+};
+
+export const THEMES = [
+  {
+    type: THEMES_TYPE.BASIC_SALES,
+    styles: basicSales,
+  },
+];